turtle_sine/src/requirements.txt
Pillow==2.7.0
numpy==1.8.0rc1
matplotlib==1.3.1
cv2==1.0
imutils==0.3.1
python==2.7